pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "commitgen"
version = "0.1.0"
description = "AI-powered Git commit message generator (Conventional Commits)"
authors = [
    { name="Yahya Jaafar" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.0",
    "google-generativeai>=0.8.3",
]

[project.scripts]
commitgen = "commitgen.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["commitgen*"]

